requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.1
Deprecated==1.2.14
discord-webhook==1.3.0
diskcache==5.6.3
disnake==2.9.1
distro==1.8.0
dm-tree==0.1.8
docutils==0.20.1
fastapi==0.104.0
feedparser==6.0.10
filelock==3.13.0
FLAML==2.1.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fluidsynth==0.2
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.5.4
git-filter-repo==2.38.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
html5lib==1.1
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
kaleido==0.2.1
keras==2.15.0
keyring==24.2.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.343
langchain-core==0.0.7
langchain-decorators==0.2.3
langchain-experimental==0.0.35
langsmith==0.0.67
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
midi2audio==0.1.1
mingus==0.6.1
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==10.1.0
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
music21==9.1.0
mypy-extensions==1.0.0
namex==0.0.7
natsort==8.4.0
nh3==0.2.14
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.1
openai==1.6.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
orjson==3.9.10
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
peewee==3.17.0
Pillow==10.1.0
pkginfo==1.9.6
plotly==5.18.0
polygon==1.1.3
polygon-api-client==1.13.3
posthog==3.0.2
priority==2.0.0
promptwatch==0.3.0
protobuf==4.23.4
psycopg2==2.9.9
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytrends==4.9.2
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
Quart==0.19.3
readme-renderer==42.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
rsa==4.9
safetensors==0.4.1
scipy==1.11.4
seaborn==0.13.0
selenium==4.14.0
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.22
starlette==0.27.0
stocksera==0.1.21
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tiktoken==0.5.1
tokenizers==0.14.1
tqdm==4.66.1
transformers==4.35.2
trio==0.22.2
trio-websocket==0.11.1
twine==4.0.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
waitress==2.1.2
watchfiles==0.21.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.32
zipp==3.17.0
